efetuar a compra dos produtos

+++ src/carts/service/carts.service.ts
@@ -1,10 +1,10 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { create } from 'domain';
+import { check } from 'prettier';
 import { ProductEntity } from 'src/products/entities/product.entity';
 import { Equal, Repository } from 'typeorm';
+import { BuyProductsDto } from '../dto/buy-products.dto';
 import { CreateCartDto } from '../dto/create-cart.dto';
 import { TotalCartDTO } from '../dto/total-cart.dto';
-import { UpdateCartDto } from '../dto/update-cart.dto';
 import { CartProductEntity } from '../entities/cart-products.entity';
 import { CartEntity } from '../entities/cart.entity';
 
@@ -19,12 +19,40 @@ export class CartsService {
     private readonly cartProductRepository: Repository<CartProductEntity>
   ) { }
 
+  async buyCart(info: BuyProductsDto) {
+    return new Promise(async (resolve, reject) => {
+      try {
+        const data = new Date(info.payment.dueDate);
+
+        const checkCard = info.payment.cardNumber === "4444 4444 4444 4444" && info.payment.cvv === "222" && data > new Date();
+        const cartIsActive = await this.cartRepository.findOne({
+          where: {
+            id: info.cartId
+          }
+        })
+
+        if (checkCard && cartIsActive.active) {
+          const closeCart = await this.cartRepository.update({ id: info.cartId }, { active: false })
+          if (closeCart.affected > 0) {
+            resolve("Compra realizada com sucesso")
+
+          }
+          resolve("Problema na compra, tente novamente")
+        }
+        resolve("Dados inválidos.")
+      } catch (error) {
+        reject(error);
+      }
+    })
+  }
+
   createCart(createCart: CreateCartDto): Promise<any> {
     return new Promise(async (resolve, reject) => {
       try {
         if (createCart.cartId === null) {
           const cart = this.cartRepository.create()
           cart.userId = createCart.userId;
+          cart.active = true;
           await this.cartRepository.save(cart);
         }
 
@@ -37,7 +65,7 @@ export class CartsService {
           where: { id: createCart.productId }
         });
 
-        if (product != null) {
+        if (product != null && cartId.active) {
           const addProductInCart = this.cartProductRepository.create()
           addProductInCart.cartId = cartId;
           addProductInCart.productId = product;
@@ -65,27 +93,30 @@ export class CartsService {
     });
   }
 
-  findAll() {
-    return `This action returns all carts`;
-  }
-
   findProductsInCart(id: number) {
     return new Promise(async (resolve, reject) => {
       try {
-        const productsInCart = await this.cartProductRepository.find({
+        const cartIsActive = await this.cartRepository.findOne({
           where: {
-            cartId: Equal(id)
+            id: id
           }
         })
-        if (productsInCart.length > 0) {
-          const cartData = new TotalCartDTO();
-          cartData.items = productsInCart;
-          const initialValue = 0;
-          cartData.total = productsInCart.reduce(
-            (accumulator, currentValue) => accumulator + currentValue.productId.price,
-            initialValue
-          );
-          resolve(cartData);
+        if (cartIsActive.active) {
+          const productsInCart = await this.cartProductRepository.find({
+            where: {
+              cartId: Equal(id)
+            }
+          })
+          if (productsInCart.length > 0) {
+            const cartData = new TotalCartDTO();
+            cartData.items = productsInCart;
+            const initialValue = 0;
+            cartData.total = productsInCart.reduce(
+              (accumulator, currentValue) => accumulator + currentValue.productId.price,
+              initialValue
+            );
+            resolve(cartData);
+          }
         }
         resolve("Sem itens no carrinho");
       } catch (error) {
@@ -94,35 +125,38 @@ export class CartsService {
     })
   }
 
-  update(id: number, updateCartDto: UpdateCartDto) {
-    return `This action updates a #${id} cart`;
-  }
-
   remove(id: number, cartId: number) {
     return new Promise(async (resolve, reject) => {
       try {
-        const removed = await this.cartProductRepository.delete({
-          productId: Equal(id),
-          cartId: Equal(cartId)
-        })
-
-        if(removed.affected === 0){
-          resolve("Erro ao remover ou item não presente no carrinho")
-        }
-        const atualCart = await this.cartProductRepository.find({
-          where:{
-            cartId: Equal(cartId),
+        const cartIsActive = await this.cartRepository.findOne({
+          where: {
+            id: id
           }
         })
-        if (atualCart.length > 0) {
-          const cartData = new TotalCartDTO();
-          cartData.items = atualCart;
-          const initialValue = 0;
-          cartData.total = atualCart.reduce(
-            (accumulator, currentValue) => accumulator + currentValue.productId.price,
-            initialValue
-          );
-          resolve(cartData);
+        if (cartIsActive.active) {
+          const removed = await this.cartProductRepository.delete({
+            productId: Equal(id),
+            cartId: Equal(cartId)
+          })
+
+          if (removed.affected === 0) {
+            resolve("Erro ao remover ou item não presente no carrinho")
+          }
+          const atualCart = await this.cartProductRepository.find({
+            where: {
+              cartId: Equal(cartId),
+            }
+          })
+          if (atualCart.length > 0) {
+            const cartData = new TotalCartDTO();
+            cartData.items = atualCart;
+            const initialValue = 0;
+            cartData.total = atualCart.reduce(
+              (accumulator, currentValue) => accumulator + currentValue.productId.price,
+              initialValue
+            );
+            resolve(cartData);
+          }
         }
         resolve("O carrinho está vazio");
       } catch (error) {
